pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "polymarket-arb-bot"
version = "0.1.0"
description = "Polymarket arb bot — rebate-maker, grid-maker, and observer strategies"
requires-python = ">=3.11"
dependencies = [
    "py-clob-client",
    "requests",
    "python-dotenv",
    "pyyaml",
    "web3",
    "websockets>=12.0",
    "sqlalchemy>=2.0",
    "pandas>=3.0.0",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
]

[project.scripts]
rebate-maker = "rebate_maker.bot:main"
observer = "observer.bot:main"
grid-maker = "grid_maker.bot:main"
clob-limit-test = "grid_maker.clob_limit_test:main"

[dependency-groups]
dev = ["ruff", "pytest"]

[tool.hatch.build.targets.wheel]
packages = ["src/shared", "src/rebate_maker", "src/observer", "src/grid_maker"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
